dev: Move RemoteControlBoard as plugin
Also cleanup
