+++ src/devices/CMakeLists.txt
@@ -70,6 +70,7 @@ yarp_begin_plugin_library(yarpmod OPTION YARP_COMPILE_DEVICE_PLUGINS
   add_subdirectory(TestMotor)
   add_subdirectory(ServerFrameGrabber)
   add_subdirectory(RemoteFrameGrabber)
+  add_subdirectory(RemoteControlBoard)
 
 
   # We can also suck in other device libraries built the same way.
