--- src/libYARP_dev/src/devices/RemoteControlBoard/RemoteControlBoard.h
@@ -1,14 +1,0 @@
-/*
- * Copyright (C) 2006-2019 Istituto Italiano di Tecnologia (IIT)
- * Copyright (C) 2006-2010 RobotCub Consortium
- * All rights reserved.
- *
- * This software may be modified and distributed under the terms of the
- * BSD-3-Clause license. See the accompanying LICENSE file for details.
- */
-
-#ifndef YARP_DEV_REMOTECONTROLBOARD_REMOTECONTROLBOARD_H
-#define YARP_DEV_REMOTECONTROLBOARD_REMOTECONTROLBOARD_H
-
-
-#endif // YARP_DEV_REMOTECONTROLBOARD_REMOTECONTROLBOARD_H
