+++ tests/libYARP_dev/ControlBoardWrapper2Test.cpp
@@ -27,6 +27,7 @@ TEST_CASE("dev::ControlBoardWrapper2", "[yarp::dev]")
 {
     YARP_REQUIRE_PLUGIN("group", "device");
     YARP_REQUIRE_PLUGIN("test_motor", "device");
+    YARP_REQUIRE_PLUGIN("remote_controlboard", "device");
 
     Network::setLocalMode(true);
